refactor: move JSONSessionStorage to a JSONStorage utils module
